Add ability to select build target to haskell-mode through shm.
This is just a stop gap. The build target is stored in a haskell-mode custome varialble.
The query to set the build target (this commit) is added to existing shm code. This is just
to get a client up and running but will be refactored into its own entitiy soon.